♻️ add equivalent helper for building writable selector families
