+++ packages/atom.io/internal/src/families/create-selector-family.ts
@@ -1,20 +1,14 @@
 import type {
-	FamilyMetadata,
 	ReadonlySelectorFamily,
 	ReadonlySelectorFamilyOptions,
 	SelectorFamily,
 	SelectorFamilyOptions,
-	SelectorToken,
 } from "atom.io"
 import type { Json } from "atom.io/json"
-import { stringifyJson } from "atom.io/json"
 
-import { newest } from "../lineage"
-import { createSelector } from "../selector"
 import type { Store } from "../store"
-import { deposit } from "../store"
-import { Subject } from "../subject"
 import { createReadonlySelectorFamily } from "./create-readonly-selector-family"
+import { createWritableSelectorFamily } from "./create-writable-selector-family"
 
 export function createSelectorFamily<T, K extends Json.Serializable>(
 	options: SelectorFamilyOptions<T, K>,
@@ -33,37 +27,5 @@ export function createSelectorFamily<T, K extends Json.Serializable>(
 	if (isReadonly) {
 		return createReadonlySelectorFamily(options, store)
 	}
-	const target = newest(store)
-	const subject = new Subject<SelectorToken<T>>()
-
-	const selectorFamily = Object.assign(
-		(key: K): SelectorToken<T> => {
-			const subKey = stringifyJson(key)
-			const family: FamilyMetadata = { key: options.key, subKey }
-			const fullKey = `${options.key}(${subKey})`
-			const existing = target.selectors.get(fullKey)
-			if (existing) {
-				return deposit(existing)
-			}
-			const token = createSelector<T>(
-				{
-					key: fullKey,
-					get: options.get(key),
-					set: options.set(key),
-				},
-				family,
-				store,
-			)
-			subject.next(token)
-			return token
-		},
-		{
-			key: options.key,
-			type: `selector_family`,
-			subject,
-			install: (store: Store) => createSelectorFamily(options, store),
-		} as const,
-	) as SelectorFamily<T, K>
-	target.families.set(options.key, selectorFamily)
-	return selectorFamily
+	return createWritableSelectorFamily(options, store)
 }
